@@ -5,7 +5,7 @@
   "license": "MIT",
   "sideEffects": false,
   "scripts": {
-    "dev": "webpack-dev-server",
+    "dev": "webpack serve",
     "dist": "NODE_ENV=production webpack-cli",
     "deploy": "npm run test && npm run dist && firebase deploy",
     "deploy_rules": "firebase deploy --only firestore",
@@ -13,48 +13,48 @@
     "upgrade-interactive": "npm-check --update"
   },
   "dependencies": {
-    "@material-ui/core": "4.9.8",
-    "@material-ui/icons": "4.9.1",
+    "@material-ui/core": "4.11.2",
+    "@material-ui/icons": "4.11.2",
     "classnames": "^2.2.6",
-    "d3-array": "2.4.0",
-    "d3-axis": "^1.0.12",
-    "d3-scale": "3.2.1",
-    "d3-selection": "1.4.1",
-    "d3-shape": "1.3.7",
-    "d3-time": "1.1.0",
-    "d3-transition": "1.3.2",
+    "d3-array": "2.9.1",
+    "d3-axis": "2.0.0",
+    "d3-scale": "3.2.3",
+    "d3-selection": "2.0.0",
+    "d3-shape": "2.0.0",
+    "d3-time": "2.0.0",
+    "d3-transition": "2.0.0",
     "firebase": "7.13.1",
-    "moment": "^2.24.0",
-    "particlesjs": "^2.2.3",
+    "moment": "2.29.1",
     "prop-types": "^15.7.2",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-firebaseui": "4.1.0",
-    "react-redux": "7.2.0",
-    "react-router-dom": "5.1.2",
+    "react-redux": "7.2.2",
+    "react-router-dom": "5.2.0",
     "redux": "4.0.5"
   },
   "devDependencies": {
-    "@babel/core": "7.9.0",
-    "@babel/preset-env": "7.9.0",
-    "@babel/preset-react": "7.9.4",
-    "babel-loader": "8.1.0",
+    "@babel/core": "7.12.10",
+    "@babel/plugin-transform-runtime": "7.12.10",
+    "@babel/preset-env": "7.12.11",
+    "@babel/preset-react": "7.12.10",
+    "babel-loader": "8.2.2",
     "clean-webpack-plugin": "3.0.0",
-    "eslint": "6.8.0",
-    "eslint-config-airbnb": "18.1.0",
-    "eslint-loader": "3.0.3",
-    "eslint-plugin-import": "2.20.2",
-    "eslint-plugin-jsx-a11y": "6.2.3",
-    "eslint-plugin-react": "7.19.0",
-    "eslint-plugin-react-hooks": "2.5.0",
-    "firebase-tools": "7.16.1",
-    "html-webpack-plugin": "4.0.3",
-    "jest": "25.2.3",
+    "eslint": "7.2.0",
+    "eslint-config-airbnb": "18.2.1",
+    "eslint-loader": "4.0.2",
+    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-jsx-a11y": "6.4.1",
+    "eslint-plugin-react": "7.21.5",
+    "eslint-plugin-react-hooks": "4.0.0",
+    "firebase-tools": "9.1.0",
+    "html-webpack-plugin": "4.5.0",
+    "jest": "26.6.3",
     "npm-check": "5.9.2",
-    "webpack": "4.42.1",
-    "webpack-cli": "3.3.11",
-    "webpack-dev-server": "3.10.3",
-    "webpack-merge": "4.2.2",
-    "webpack-notifier": "1.8.0"
+    "webpack": "5.11.1",
+    "webpack-cli": "4.3.0",
+    "webpack-dev-server": "3.11.0",
+    "webpack-merge": "5.7.3",
+    "webpack-notifier": "1.12.0"
   }
 }
