@@ -25,10 +25,6 @@ class Edit extends React.Component {
     this.handleEdit = this.handleEdit.bind(this);
   }
 
-  onNewValue(type, value) {
-    this.setState({ [type]: value });
-  }
-
   handleEdit() {
     const {
       name,
@@ -65,6 +61,10 @@ class Edit extends React.Component {
     }));
 
     return true;
+  }
+
+  onNewValue(type, value) {
+    this.setState({ [type]: value });
   }
 
   render() {
