@@ -30,10 +30,6 @@ function getCurrentUser() {
   return currentUser;
 }
 
-// function onAuthStateChanged(callback) {
-//   return firebase.auth().onAuthStateChanged(callback);
-// }
-
 function signOut() {
   return firebase.auth().signOut();
 }
@@ -41,6 +37,6 @@ function signOut() {
 export {
   requiresAuth,
   getCurrentUser,
-  // onAuthStateChanged,
+  isLoggedIn,
   signOut,
 };
